backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
openpyxl
python-multipart
pydantic
numpy
lxml
requests